fix ckpt downloading bug
