+++ examples/resnet152.py
@@ -4,7 +4,7 @@ import argparse
 from utils import create_onedrive_directdownload, download_rsc
 
 parser = argparse.ArgumentParser()
-parser.add_argument("root", default="", type=str)
+parser.add_argument("--path", default="", type=str)
 
 MODEL_NAME = "resnet152_baseline.pth.tar"
 ONEDRIVE_SHARED_LINK = "https://1drv.ms/u/s!Al5BF1i8TRVbiQt2whUMwlzTn69R?e=qMr3XT"
@@ -19,6 +19,7 @@ def download(model_path):
 if __name__ == '__main__':
     args = parser.parse_args()
 
-    model_path = os.join(args.root, MODEL_NAME)
+    model_path = os.path.join(args.path, MODEL_NAME)
+    model_path = os.path.abspath(model_path)
 
     download(model_path)
